Refactor; Add module pattern for RecentProductView
