+++ client/public/js/RecentProductView.js
@@ -1,0 +1,41 @@
+import CustomDomAPI from "../utils/CustomDomAPI";
+
+const RecentProductView = (function(){
+    const recentProductsEle = CustomDomAPI.querySelector('.recent-products-panel');
+    const recentProductsButton = CustomDomAPI.querySelector(".main-navbar--right__recent-products");
+    const CLASSNAME_VISIBLE = "visible"
+    
+    function init() {
+        recentProductsButton.addEventListener('mouseenter', () => {
+            render();
+            recentProductsEle.classList.add(CLASSNAME_VISIBLE);    
+        })
+        
+        recentProductsButton.addEventListener('mouseleave', () => {
+            recentProductsEle.classList.remove(CLASSNAME_VISIBLE);
+        })
+        
+        recentProductsEle.addEventListener('mouseenter', () => {
+            render();
+            recentProductsEle.classList.add(CLASSNAME_VISIBLE);
+        })
+        
+        recentProductsEle.addEventListener('mouseleave', () => {
+            recentProductsEle.classList.remove(CLASSNAME_VISIBLE);
+        })
+    }
+
+    function render () {
+        const recentProductsStr = localStorage.getItem("recent-items");
+        const recentProducts = recentProductsStr ? JSON.parse(recentProductsStr) : [];
+        let innerHTML = "";
+        for (const product of recentProducts) {
+            const productsHTML = `<img src="${product['url']}"/>`;
+            innerHTML = innerHTML + productsHTML;
+        }
+        recentProductsEle.innerHTML = innerHTML;
+    }
+    return {init};
+})();
+
+export default RecentProductView;
